- emphasized api of model in separate folder
